+++ app/src/main/java/com/survivalcoding/runningtracker/domain/model/Run.kt
@@ -2,7 +2,7 @@ package com.survivalcoding.runningtracker.domain.model
 
 data class Run(
     val id: Int? = null,
-    val distanceInMeters: Int = 0,
+    val distanceInMeters: Double = 0.0,
     val timeInMillis: Long = 0L,
     val timestamp: Long = 0L,
     val avgSpeedInKMH: Float = 0f,
